requirements.txt
# ------------------------------------------------------------ #
#                            Others                            #
# ------------------------------------------------------------ #

pyyaml
matplotlib

# ------------------------------------------------------------ #
#                        Data Preprocess                       #
# ------------------------------------------------------------ #

numpy
pandas
scikit-learn

# ------------------------------------------------------------ #
#                            Pytorch                           #
# ------------------------------------------------------------ #

# Uncomment the following line if you want to use GPU.
-f https://download.pytorch.org/whl/cu117
torch
-f https://download.pytorch.org/whl/cu117
torchtext